requirements.txt
pdfplumber
sentence-transformers
faiss-cpu
rank-bm25
flask
sqlite-utils
numpy
huggingface_hub
hf_xet
scikit-learn
joblib